Add DMG packaging script for distribution
This script creates a distributable DMG file for easy installation:

- Packages the built app into a compressed DMG image
- Creates an Applications symlink for drag-and-drop installation
- Includes a README.txt with installation instructions and requirements
- Uses hdiutil to create a compressed UDZO format DMG
- Outputs file size and location information

The DMG provides a professional installation experience where users can
simply drag the app to their Applications folder. This is the standard
distribution method for macOS applications and makes installation
intuitive for end users.

Version is set to 1.0.0 for the initial release.